refactor: Dto 타입 변경에 따른 Controller / Assembler 구조 수정
