--- src/main/java/com/commerce/backendserver/product/application/ProductService.java
@@ -1,12 +1,0 @@
-package com.commerce.backendserver.product.application;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@Transactional
-@RequiredArgsConstructor
-public class ProductService {
-
-}
